fix(vagrant): Checks if booting already provisioned box
If the box already has a the vim taglist plugin installed
do not attempt a second install.

+++ vagrant/provision.sh
@@ -11,10 +11,15 @@ sudo apt-get install --yes unzip exuberant-ctags
 # TODO: Set up vim plugins
 sudo apt-get install --yes vim
 # Taglist plugin
-wget -O /tmp/taglist_46.zip  http://garr.dl.sourceforge.net/project/vim-taglist/vim-taglist/4.6/taglist_46.zip
-mkdir -p ~/.vim
-unzip /tmp/taglist_46.zip -d ~/.vim
-echo "filetype on" >> ~/.vimrc
+if [ -f ~/.vim/plugin/taglist.vim ];
+then
+	echo "vim taglist already installed"
+else
+	wget -O /tmp/taglist_46.zip  http://garr.dl.sourceforge.net/project/vim-taglist/vim-taglist/4.6/taglist_46.zip
+	mkdir -p ~/.vim
+	unzip -u /tmp/taglist_46.zip -d ~/.vim
+	echo "filetype on" >> ~/.vimrc
+fi
 
 # Install and run node.js
 # The jasmine-node will not finish but continue to wath and test files in node.
